requirements.txt
flask
flask-sqlalchemy
flask-login
flask-wtf
flask-bootstrap
sqlalchemy
Werkzeug
python-dotenv